requirements.txt
langchain_community 
langchain
langchain_core
streamlit 
python-dotenv
langchain_chroma
chromadb
